requirements.txt
absl-py==1.1.0
androguard==3.4.0a1
appdirs==1.4.4
apturl==0.5.2
arandr==0.1.10
argcomplete==1.8.1
asn1crypto==1.4.0
async-generator==1.10
attrs==21.2.0
autopep8==1.6.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
binwalk==2.3.3
blinker==1.4
borgbackup==1.2.0
Brlapi==0.8.3
Brotli==1.0.9
cached-property==1.5.2
cairocffi==1.2.0
canadian-ham-exam==1.0.0
Cartopy==0.0.0
certifi==2020.6.20
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.6
chex==0.1.3
chirp==0.3.0.dev0
chrome-gnome-shell==0.0.0
click==8.0.3
click-plugins==1.1.1
colorama==0.4.4
coloredlogs==7.3
command-not-found==0.3
commonmark==0.9.1
conda-package-handling==1.7.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
Cython==0.29.32
dbus-next==0.2.3
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
diffoscope==205
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
dm-tree==0.1.7
dnspython==2.1.0
docker==5.0.3
docker-compose==1.29.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
EditorConfig==0.12.2
enjarify==1.0.3
etils==0.6.0
evdev==1.4.0
fasteners==0.14.1
filelock==3.6.0
flatbuffers==2.0
flax==0.4.2
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.5.2
gitdb==4.0.9
GitPython==3.1.27
gyp==0.1
html5lib==1.1
httpie==2.6.0
httplib2==0.20.2
humanfriendly==10.0
idna==3.3
img2pdf==0.4.2
importlib-metadata==4.6.4
importlib-resources==5.8.0
input-remapper==1.4.0
install==1.3.5
iotop==0.6
ipython==7.31.1
jax==0.2.22
jaxlib==0.3.14
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.3
jsbeautifier==1.14.0
jsondiff==1.3.1
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libarchive-c==2.9
llfuse==1.3.8
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mechanize==0.4.7
membernator==1.1.0
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
msgpack==1.0.3
mypy==0.971
mypy-extensions==0.4.3
netifaces==0.11.0
networkx==2.4
notify2==0.3
numpy==1.21.5
oauthlib==3.2.0
ocrmypdf==13.4.0+dfsg
olefile==0.46
opencv-python==4.6.0.66
opt-einsum==3.3.0
optax==0.1.2
outcome==1.1.0
packaging==21.3
paramiko==2.9.3
parso==0.8.1
pathtools==0.1.2
pdfminer.six===-VERSION-
pexpect==4.8.0
pickleshare==0.7.5
pikepdf==5.0.1+dfsg
Pillow==9.0.1
platformdirs==2.5.1
pluggy==0.13.0
ply==3.11
progressbar==2.5
promise==2.3
prompt-toolkit==3.0.28
protobuf==3.12.4
proton-client==0.7.1
protonvpn-cli==3.13.0
protonvpn-gui==1.11.0
protonvpn-nm-lib==3.13.0
psutil==5.9.0
ptyprocess==0.7.0
pudb==2020.1
py==1.10.0
pycairo==1.20.1
pycodestyle==2.8.0
pycparser==2.21
pycups==2.0.1
pydbus==0.6.0
pydot==1.4.2
pyfuse3==3.2.0
pygame==2.1.2
pygccxml==2.2.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
PyOpenGL==3.1.5
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyperclip==1.8.2
pyproj==3.3.0
PyQSO==1.1.0
PyQt-Qwt==1.2.2
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyqtgraph==0.12.4
pyRFC3339==1.1
pyrsistent==0.18.1
pyserial==3.5
pyshp==2.2.0
PySocks==1.7.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==0.19.2
python-gnupg==0.4.8
python-magic==0.4.24
pythondialog==3.5.1
pythran==0.10.0
pytz==2022.1
pyusb==1.2.1.post1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
pyzmq==22.3.0
qtile @ file:///home/fortyseven/opt/qtile
quisk==4.1.93
reportlab==3.6.8
requests==2.25.1
requests-toolbelt==0.9.1
rich==11.2.0
rpm==4.17.0
scapy==2.4.4
scipy==1.8.0
scour==0.38.2
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sentry-sdk==1.6.0
setproctitle==1.2.2
Shapely==1.8.0
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.1.0
soupsieve==2.3.1
srt==3.5.2
ssh-audit==2.5.0
ssh-import-id==5.11
stepic==0.5.0
sympy==1.9
systemd-python==234
texttable==1.6.4
thrift==0.16.0
tlsh==0.2.0
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
torch==1.12.0
tornado==6.1
tqdm==4.57.0
traitlets==5.1.1
trio==0.19.0
typing_extensions==4.3.0
ubuntu-advantage-tools==27.10
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.5
urwid==2.1.2
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.13.0+ds
vosk==0.3.38
wadllib==1.3.6
wandb==0.12.20
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
wxPython==4.0.7
xcffib==0.11.1
xdg==5
xkit==0.0.0
xlrd==1.2.0
yt-dlp==2022.8.19
zipp==1.0.0
